src/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
motor==3.7.1
passlib[bcrypt]
python-jose[cryptography]
python-multipart
celery==5.3.6
redis==5.0.1
openai>=1.0.0
python-docx==1.1.0
pypdf==3.17.4
sendgrid
